Finish moving the rest of the CRUD stuff over to CrudModule
