+++ CrudModule/GraphQL/Types/SpeciesGraphType.cs
@@ -1,22 +1,24 @@
 ﻿using DataAccess;
 using DataAccess.Constants;
+using DataAccessGraphQL;
+using DataAccessGraphQL.GraphTypes;
 using GraphQL.Types;
 
-namespace DataAccessGraphQL.GraphTypes
+namespace CrudModule.GraphQL.Types
 {
-    public class ContactGraphType : ObjectGraphType<Contact>
+    public class SpeciesGraphType : ObjectGraphType<Species>
     {
-        public ContactGraphType(GraphQLUserContextAccessor contextAccessor)
+        public SpeciesGraphType(GraphQLUserContextAccessor contextAccessor)
         {
             var context = contextAccessor.Context;
 
-            this.Field<StringGraphType>("id", resolve: ctx => ctx.Source.ContactId);
+            this.Field<StringGraphType>("id", resolve: ctx => ctx.Source.SpeciesId);
 
             this.FieldAsync<StringGraphType>(
                 "name",
                 resolve: async ctx =>
                 {
-                    await context.EnforceHasPolicyAsync(Permissions.Contact.Read);
+                    await context.EnforceHasPolicyAsync(Permissions.Species.Read);
                     return ctx.Source.Name;
                 }
             );
@@ -25,7 +27,7 @@ namespace DataAccessGraphQL.GraphTypes
                 "notes",
                 resolve: async ctx =>
                 {
-                    await context.EnforceHasPolicyAsync(Permissions.Contact.ReadNotes);
+                    await context.EnforceHasPolicyAsync(Permissions.Species.ReadNotes);
                     return ctx.Source.NoteOwner?.NoteEntries;
                 }
             );
