+++ DataAccessGraphQL/DataAccessGraphQL.csproj
@@ -17,4 +17,8 @@
     <ProjectReference Include="..\DataAccess\DataAccess.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Mutations\" />
+  </ItemGroup>
+
 </Project>
